+++ sdk/ml/azure-ai-ml/azure/ai/ml/entities/_workspace/networking.py
@@ -72,7 +72,6 @@ class OutboundRule(ABC):
         return None
 
 
-@experimental
 class FqdnDestination(OutboundRule):
     """Class representing a FQDN outbound rule.
 
@@ -100,7 +99,6 @@ class FqdnDestination(OutboundRule):
         }
 
 
-@experimental
 class PrivateEndpointDestination(OutboundRule):
     """Class representing a Private Endpoint outbound rule.
 
@@ -153,7 +151,6 @@ class PrivateEndpointDestination(OutboundRule):
         }
 
 
-@experimental
 class ServiceTagDestination(OutboundRule):
     """Class representing a Service Tag outbound rule.
 
@@ -205,7 +202,6 @@ class ServiceTagDestination(OutboundRule):
         }
 
 
-@experimental
 class ManagedNetwork:
     def __init__(
         self,
@@ -248,7 +244,6 @@ class ManagedNetwork:
         )
 
 
-@experimental
 class ManagedNetworkProvisionStatus:
     """ManagedNetworkProvisionStatus.